# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gavin API",
    "description": "Content publishing and administration API for Gavin.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/articles": {
      "get": {
        "tags": [
          "admin articles"
        ],
        "summary": "List Articles",
        "operationId": "list_articles_api_v1_admin_articles_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleResponse"
                  },
                  "title": "Response List Articles Api V1 Admin Articles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin articles"
        ],
        "summary": "Create Article",
        "operationId": "create_article_api_v1_admin_articles_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/articles/{article_id}": {
      "get": {
        "tags": [
          "admin articles"
        ],
        "summary": "Get Article",
        "operationId": "get_article_api_v1_admin_articles__article_id__get",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin articles"
        ],
        "summary": "Update Article",
        "operationId": "update_article_api_v1_admin_articles__article_id__patch",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Article Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin articles"
        ],
        "summary": "Trash Article",
        "operationId": "trash_article_api_v1_admin_articles__article_id__delete",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/articles": {
      "get": {
        "tags": [
          "public articles"
        ],
        "summary": "List Public Articles",
        "operationId": "list_public_articles_api_v1_articles_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicArticle"
                  },
                  "title": "Response List Public Articles Api V1 Articles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/{year}/{month}/{slug}": {
      "get": {
        "tags": [
          "public articles"
        ],
        "summary": "Get Public Article",
        "operationId": "get_public_article_api_v1_articles__year___month___slug__get",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Month"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicArticle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/{year}/{month}/{slug}/context": {
      "get": {
        "tags": [
          "public articles"
        ],
        "summary": "Get Public Article Context",
        "operationId": "get_public_article_context_api_v1_articles__year___month___slug__context_get",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Month"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleContext"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArticleContext": {
        "properties": {
          "previous": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RelatedArticle"
              },
              {
                "type": "null"
              }
            ]
          },
          "next": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RelatedArticle"
              },
              {
                "type": "null"
              }
            ]
          },
          "related": {
            "items": {
              "$ref": "#/components/schemas/RelatedArticle"
            },
            "type": "array",
            "title": "Related"
          },
          "backlinks": {
            "items": {
              "$ref": "#/components/schemas/PublicBacklink"
            },
            "type": "array",
            "title": "Backlinks"
          }
        },
        "type": "object",
        "title": "ArticleContext"
      },
      "ArticleCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 180,
            "minLength": 1,
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "summary": {
            "type": "string",
            "maxLength": 320,
            "title": "Summary",
            "default": ""
          },
          "content": {
            "type": "string",
            "title": "Content",
            "default": ""
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "tag_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Tag Ids"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/ArticleReferenceInput"
            },
            "type": "array",
            "maxItems": 20,
            "title": "References"
          }
        },
        "type": "object",
        "required": [
          "title",
          "slug"
        ],
        "title": "ArticleCreate"
      },
...
      "ArticleReference": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "external",
              "internal"
            ],
            "title": "Kind"
          },
          "display_title": {
            "type": "string",
            "title": "Display Title"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "target_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "article",
                  "project",
                  "book"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Type"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "display_title"
        ],
        "title": "ArticleReference"
      },
...
      "ArticleResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "current_revision_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Revision Id"
          },
          "current_publish_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Publish Revision"
          },
          "has_unpublished_changes": {
            "type": "boolean",
            "title": "Has Unpublished Changes",
            "default": false
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "public_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Path"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaxonomyItem"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TaxonomyItem"
            },
            "type": "array",
            "title": "Tags"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/ArticleReference"
            },
            "type": "array",
            "title": "References"
          },
          "wikilinks": {
            "items": {
              "$ref": "#/components/schemas/WikiResolution"
            },
            "type": "array",
            "title": "Wikilinks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "summary",
          "content",
          "status",
          "version",
          "published_at",
          "created_at",
          "updated_at",
          "deleted_at",
          "category",
          "tags"
        ],
        "title": "ArticleResponse"
      },
...
      "ArticleUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 180,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 320
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "tag_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Ids"
          },
          "references": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ArticleReferenceInput"
                },
                "type": "array",
                "maxItems": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "References"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "version"
        ],
        "title": "ArticleUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicArticle": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "public_path": {
            "type": "string",
            "title": "Public Path"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaxonomyItem"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TaxonomyItem"
            },
            "type": "array",
            "title": "Tags"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/PublicReference"
            },
            "type": "array",
            "title": "References"
          },
          "wikilinks": {
            "items": {
              "$ref": "#/components/schemas/WikiResolution"
            },
            "type": "array",
            "title": "Wikilinks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "summary",
          "content",
          "published_at",
          "updated_at",
          "public_path",
          "category",
          "tags"
        ],
        "title": "PublicArticle"
      },
...
      "PublicReference": {
        "properties": {
          "display_title": {
            "type": "string",
            "title": "Display Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "display_title",
          "url"
        ],
        "title": "PublicReference"
      },
...
      "TaxonomyItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "description"
        ],
        "title": "TaxonomyItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WikiResolution": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "display_title": {
            "type": "string",
            "title": "Display Title"
          },
          "public_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Path"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "article",
                  "project",
                  "book"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          }
        },
        "type": "object",
        "required": [
          "token",
          "display_title"
        ],
        "title": "WikiResolution"
      }
    }
  }
}